requirements.txt
Flask==3.0.3 
pymongo==4.10.1 
bcrypt==4.2.0 
pyjwt==2.8.0 
python-dotenv==1.0.1
flask-cors==5.0.0